Update error.html.twig
I need to test this.
The logo of an error page should not redirect the user to the operator admin area but to the user login area.
Actually UVdesk logo on error page redirect the user to the operator login so user will get always login error. This should be fixed.